requirements.txt
aerich==0.7.1
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
choicelib==0.1.5
click==8.1.3
colorama==0.4.6
dictdiffer==0.9.0
frozenlist==1.3.3
idna==3.4
iso8601==1.1.0
loguru==0.6.0
multidict==6.0.4
Pillow==9.4.0
pydantic==1.10.4
pypika-tortoise==0.1.6
python-dotenv==0.21.0
pytz==2022.7
sniffio==1.3.0
tomlkit==0.11.6
tortoise-orm==0.19.2
typing_extensions==4.4.0
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.13
watchfiles==0.18.1
win32-setctime==1.1.0
yarl==1.8.2
